fix for captcha cleanup Update to phpDOC comments in class-album.php,class-image.php, functions.php New docs for class-gallery.php
git-svn-id: http://www.zenphoto.org/svn/trunk@1032 46129fc4-274e-7f48-8cfa-5da4c607681c

+++ zp-core/controller.php
@@ -38,10 +38,12 @@ if (getOption('Use_Captcha')) {
   if ($d > $expire) {
     chdir(SERVERCACHE . "/");
     $filelist = glob('code_*.png');
-    foreach ($filelist as $file) {
-      $file = SERVERCACHE . "/" . $file;
-      if (filemtime($file) < $expire) {
-        unlink($file);
+    if ($filelist) {
+      foreach ($filelist as $file) {
+        $file = SERVERCACHE . "/" . $file;
+        if (filemtime($file) < $expire) {
+          unlink($file);
+        }
       }
     }
   }
